# Meizon-Inventory-Swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Inventory Management Service",
    "version": "1.0.1",
    "title": "Meizon Inventory Service",
    "license": {
      "name": "Apache Licence Version 2.0",
      "url": "https://opensource.org/licenses/Apache-2.0"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
...
    "/api/v1/daily/readings/location/{locationId}": {
      "get": {
        "tags": [
          "daily-reading-controller-v-1"
        ],
        "summary": "getDailyReadingById",
        "operationId": "getDailyReadingByIdUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "endDate",
            "in": "query",
            "description": "endDate",
            "required": false,
            "type": "string"
          },
          {
            "name": "locationId",
            "in": "path",
            "description": "locationId",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": false,
            "type": "integer",
            "default": 0,
            "format": "int32"
          },
          {
            "name": "size",
            "in": "query",
            "description": "size",
            "required": false,
            "type": "integer",
            "default": 10,
            "format": "int32"
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "startDate",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "daily-reading-controller-v-1"
        ],
        "summary": "addNewDailyReading",
        "operationId": "addNewDailyReadingUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "dailyReadingDto",
            "description": "dailyReadingDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DailyReadingDto"
            }
          },
          {
            "name": "locationId",
            "in": "path",
            "description": "locationId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
    "/api/v1/location/": {
      "get": {
        "tags": [
          "location-controller-v-1"
        ],
        "summary": "getAllLocations",
        "operationId": "getAllLocationsUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": false,
            "type": "integer",
            "default": 0,
            "format": "int32"
          },
          {
            "name": "size",
            "in": "query",
            "description": "size",
            "required": false,
            "type": "integer",
            "default": 10,
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "location-controller-v-1"
        ],
        "summary": "addNewLocation",
        "operationId": "addNewLocationUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "locationDto",
            "description": "locationDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LocationDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/location/cities": {
      "get": {
        "tags": [
          "location-controller-v-1"
        ],
        "summary": "getAllCities",
        "operationId": "getAllCitiesUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/location/location-types": {
      "get": {
        "tags": [
          "location-controller-v-1"
        ],
        "summary": "getAllLocationTypes",
        "operationId": "getAllLocationTypesUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/location/regions": {
      "get": {
        "tags": [
          "location-controller-v-1"
        ],
        "summary": "getAllRegions",
        "operationId": "getAllRegionsUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/location/type/{type}": {
      "get": {
        "tags": [
          "location-controller-v-1"
        ],
        "summary": "getAllLocationByType",
        "operationId": "getAllLocationByTypeUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": false,
            "type": "integer",
            "default": 0,
            "format": "int32"
          },
          {
            "name": "size",
            "in": "query",
            "description": "size",
            "required": false,
            "type": "integer",
            "default": 10,
            "format": "int32"
          },
          {
            "name": "type",
            "in": "path",
            "description": "type",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/location/{locationId}": {
      "get": {
        "tags": [
          "location-controller-v-1"
        ],
        "summary": "getLocationByLocationId",
        "operationId": "getLocationByLocationIdUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "locationId",
            "in": "path",
            "description": "locationId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [
          "location-controller-v-1"
        ],
        "summary": "editLocation",
        "operationId": "editLocationUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "locationDto",
            "description": "locationDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LocationDto"
            }
          },
          {
            "name": "locationId",
            "in": "path",
            "description": "locationId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "location-controller-v-1"
        ],
        "summary": "deleteByLocationId",
        "operationId": "deleteByLocationIdUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "locationId",
            "in": "path",
            "description": "locationId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
...
    "/api/v1/payment/{paymentId}": {
      "get": {
        "tags": [
          "payment-controller-v-1"
        ],
        "summary": "getPaymentById",
        "operationId": "getPaymentByIdUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "paymentId",
            "in": "path",
            "description": "paymentId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "payment-controller-v-1"
        ],
        "summary": "deleteByLocationId",
        "operationId": "deleteByLocationIdUsingDELETE_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "paymentId",
            "in": "path",
            "description": "paymentId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
    "AddressDto": {
      "type": "object",
      "properties": {
        "street": {
          "type": "string"
        },
        "unitNumber": {
          "type": "string"
        }
      },
      "title": "AddressDto"
    },
...
    "DailyReadingDto": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "dailyReadingId": {
          "type": "string"
        },
        "dispensedVolume": {
          "type": "integer",
          "format": "int64"
        },
        "user": {
          "$ref": "#/definitions/UserDto"
        }
      },
      "title": "DailyReadingDto"
    },
    "LocationDto": {
      "type": "object",
      "properties": {
        "address": {
          "$ref": "#/definitions/AddressDto"
        },
        "description": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "province": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      },
      "title": "LocationDto"
    },
...
    "UserDto": {
      "type": "object",
      "properties": {
        "address": {
          "$ref": "#/definitions/AddressDto"
        },
        "email": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "locationId": {
          "type": "integer",
          "format": "int64"
        },
        "mobile": {
          "type": "integer",
          "format": "int64"
        },
        "roles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserRoleDto"
          }
        },
        "userId": {
          "type": "string"
        }
      },
      "title": "UserDto"
    },
    "UserPermissionDto": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      },
      "title": "UserPermissionDto"
    },
    "UserRoleDto": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "permissions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserPermissionDto"
          }
        }
      },
      "title": "UserRoleDto"
    }
  }
}